fixed bugs around is pattern matching
