+++ src/Buckle/CommandLine/DllImportHelper.cs
@@ -2,6 +2,9 @@ using System;
 using System.IO;
 
 public static class DllImportHelper {
+    /// <summary>
+    /// Extracts libraries embedded in certain release mode builds into the installation directory
+    /// </summary>
     public static void ExtractAndLoadDlls() {
         ExtractToBaseDirectory("Resources.Belte.Runtime.dll", "Belte.Runtime.dll");
         ExtractToBaseDirectory("Resources.Belte.Graphics.dll", "Belte.Graphics.dll");
