+++ src/Buckle/Compiler/CodeAnalysis/Binding/Conversions/Conversions.cs
@@ -1,266 +1,18 @@
-using System;
 using System.Collections.Immutable;
-using System.Linq;
 using Buckle.CodeAnalysis.Symbols;
 using Buckle.CodeAnalysis.Syntax;
 using Buckle.Diagnostics;
-using Buckle.Libraries;
-using Buckle.Utilities;
-using Microsoft.CodeAnalysis.PooledObjects;
 
 namespace Buckle.CodeAnalysis.Binding;
 
-internal sealed partial class Conversions {
+internal sealed class Conversions : ConversionsBase {
     private readonly Binder _binder;
 
     internal Conversions(Binder binder) {
         _binder = binder;
     }
 
-    internal static ListExpressionTypeKind GetListExpressionTypeKind(
-        TypeSymbol destination,
-        out TypeWithAnnotations elementType) {
-        if (destination.StrippedType() is ArrayTypeSymbol arrayType) {
-            if (arrayType.isSZArray) {
-                elementType = arrayType.elementTypeWithAnnotations;
-                return ListExpressionTypeKind.Array;
-            } else {
-                elementType = new TypeWithAnnotations(
-                    ArrayTypeSymbol.CreateArray(arrayType.elementTypeWithAnnotations, arrayType.rank - 1)
-                );
-
-                return ListExpressionTypeKind.Array;
-            }
-        }
-
-        elementType = null;
-        return ListExpressionTypeKind.None;
-    }
-
-    internal Conversion ClassifyImplicitUserDefinedConversionForSwitchType(
-        TypeSymbol sourceType,
-        out TypeSymbol switchGoverningType) {
-        var result = AnalyzeImplicitUserDefinedConversionForSwitchGoverningType(sourceType);
-
-        if (result.kind == UserDefinedConversionResultKind.Valid) {
-            var analysis = result.results[result.best];
-            switchGoverningType = analysis.toType;
-        } else {
-            switchGoverningType = null;
-        }
-
-        return new Conversion(result, isImplicit: true);
-    }
-
-    internal bool HasIdentityOrImplicitReferenceConversion(TypeSymbol source, TypeSymbol destination) {
-        if (HasIdentityConversionInternal(source, destination))
-            return true;
-
-        return HasImplicitReferenceConversion(source, destination);
-    }
-
-    internal bool HasImplicitReferenceConversion(TypeSymbol source, TypeSymbol destination) {
-        if (!source.isReferenceType && !source.IsArray())
-            return false;
-
-        if (destination.specialType == SpecialType.Object)
-            return true;
-
-        switch (source.typeKind) {
-            case TypeKind.Class:
-                if (destination.IsClassType() && Conversion.IsBaseClass(source, destination))
-                    return true;
-
-                return HasImplicitConversionToInterface(source, destination);
-            case TypeKind.Interface:
-                return HasImplicitConversionToInterface(source, destination);
-            case TypeKind.TemplateParameter:
-                return HasImplicitReferenceTemplateParameterConversion((TemplateParameterSymbol)source, destination);
-            case TypeKind.Array:
-                // TODO Again, we probably don't want covariance
-                // return HasImplicitConversionFromArray(source, destination);
-                return false;
-        }
-
-        return false;
-    }
-
-    private bool HasImplicitReferenceTemplateParameterConversion(
-        TemplateParameterSymbol source,
-        TypeSymbol destination) {
-        if (source.isValueType)
-            return false;
-
-        // if (source.AllowsRefLikeType) {
-        //     return false;
-        // }
-
-        if (HasImplicitEffectiveBaseConversion(source, destination))
-            return true;
-
-        if (HasImplicitEffectiveInterfaceSetConversion(source, destination))
-            return true;
-
-        if (destination is TemplateParameterSymbol { allowsRefLikeType: false } &&
-            source.DependsOn((TemplateParameterSymbol)destination)) {
-            return true;
-        }
-
-        return false;
-    }
-
-    private bool HasImplicitEffectiveInterfaceSetConversion(TemplateParameterSymbol source, TypeSymbol destination) {
-        return HasVarianceCompatibleInterfaceInEffectiveInterfaceSet(source, destination);
-    }
-
-    private bool HasVarianceCompatibleInterfaceInEffectiveInterfaceSet(
-        TemplateParameterSymbol source,
-        TypeSymbol destination) {
-        if (!destination.IsInterfaceType())
-            return false;
-
-        foreach (var i in source.allEffectiveInterfaces) {
-            if (HasInterfaceVarianceConversion(i, destination))
-                return true;
-        }
-
-        return false;
-    }
-
-    private bool HasImplicitEffectiveBaseConversion(TemplateParameterSymbol source, TypeSymbol destination) {
-        var effectiveBaseClass = source.effectiveBaseClass;
-
-        if (HasIdentityConversionInternal(effectiveBaseClass, destination))
-            return true;
-
-        if (IsBaseClass(effectiveBaseClass, destination))
-            return true;
-
-        if (HasAnyBaseInterfaceConversion(effectiveBaseClass, destination))
-            return true;
-
-        return false;
-    }
-
-    internal bool IsBaseClass(TypeSymbol derivedType, TypeSymbol baseType) {
-        if (!baseType.IsClassType())
-            return false;
-
-        for (TypeSymbol b = derivedType.baseType; b is not null; b = b.baseType) {
-            if (HasIdentityConversionInternal(b, baseType))
-                return true;
-        }
-
-        return false;
-    }
-
-    private bool HasImplicitConversionToInterface(TypeSymbol source, TypeSymbol destination) {
-        if (!destination.IsInterfaceType())
-            return false;
-
-        if (source.IsClassType())
-            return HasAnyBaseInterfaceConversion(source, destination);
-
-        if (source.IsInterfaceType()) {
-            if (HasAnyBaseInterfaceConversion(source, destination))
-                return true;
-
-            if (!HasIdentityConversionInternal(source, destination) &&
-                HasInterfaceVarianceConversion(source, destination)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private bool HasAnyBaseInterfaceConversion(TypeSymbol derivedType, TypeSymbol baseType) {
-        return ImplementsVarianceCompatibleInterface(derivedType, baseType);
-    }
-
-    private bool ImplementsVarianceCompatibleInterface(TypeSymbol derivedType, TypeSymbol baseType) {
-        if (!baseType.IsInterfaceType())
-            return false;
-
-        if (derivedType is not NamedTypeSymbol d)
-            return false;
-
-        foreach (var i in d.allInterfaces) {
-            if (HasInterfaceVarianceConversion(i, baseType))
-                return true;
-        }
-
-        return false;
-    }
-
-    private bool HasInterfaceVarianceConversion(TypeSymbol source, TypeSymbol destination) {
-        if (source is not NamedTypeSymbol s || destination is not NamedTypeSymbol d)
-            return false;
-
-        if (!s.IsInterfaceType() || !d.IsInterfaceType())
-            return false;
-
-        return HasVariantConversion(s, d);
-    }
-
-    private bool HasVariantConversion(NamedTypeSymbol source, NamedTypeSymbol destination) {
-        // TODO Recursion guard
-        // if (currentRecursionDepth >= MaximumRecursionDepth)
-        //     return false;
-
-        var quickResult = HasVariantConversionQuick(source, destination);
-
-        if (quickResult.HasValue())
-            return quickResult.Value();
-
-        // return this.CreateInstance(currentRecursionDepth + 1).
-        return HasVariantConversionNoCycleCheck(source, destination);
-    }
-
-    private bool HasVariantConversionNoCycleCheck(NamedTypeSymbol source, NamedTypeSymbol destination) {
-        // TODO We probably don't want covariance or contravariance, but this is where we would do some checks for it
-        return true;
-    }
-
-    private ThreeState HasVariantConversionQuick(NamedTypeSymbol source, NamedTypeSymbol destination) {
-        if (HasIdentityConversionInternal(source, destination))
-            return ThreeState.True;
-
-        var typeSymbol = source.originalDefinition;
-
-        if (!TypeSymbol.Equals(typeSymbol, destination.originalDefinition, TypeCompareKind.ConsiderEverything))
-            return ThreeState.False;
-
-        return ThreeState.Unknown;
-    }
-
-    private static bool HasIdentityConversionInternal(TypeSymbol type1, TypeSymbol type2) {
-        return type1.Equals(type2, TypeCompareKind.AllIgnoreOptions);
-    }
-
-    private Conversion GetImplicitNullptrExpressionConversion(
-        BoundUnconvertedNullptrExpression ptrExpression,
-        TypeSymbol destination) {
-        var ptrExpressionConversion = GetNullptrExpressionConversion(ptrExpression, destination);
-
-        if (ptrExpressionConversion.exists)
-            return ptrExpressionConversion;
-
-        return Conversion.None;
-    }
-
-    private Conversion GetImplicitEnumFieldExpressionConversion(
-        BoundUnconvertedImplicitEnumFieldExpression fieldAccess,
-        TypeSymbol destination) {
-        var fieldAccessConversion = GetEnumFieldExpressionConversion(fieldAccess, destination);
-
-        if (fieldAccessConversion.exists)
-            return fieldAccessConversion;
-
-        return Conversion.None;
-    }
-
-    private Conversion GetImplicitExtendedLiteralExpressionConversion(
+    internal override Conversion GetImplicitExtendedLiteralExpressionConversion(
         BoundUnconvertedExtendedLiteralExpression extended,
         TypeSymbol destination) {
         var extendedConversion = GetExtendedLiteralExpressionConversion(_binder, extended, destination);
@@ -271,607 +23,7 @@ internal sealed partial class Conversions {
         return Conversion.None;
     }
 
-    private Conversion GetImplicitListExpressionConversion(
-        BoundUnconvertedInitializerList listExpression,
-        TypeSymbol destination) {
-        var listExpressionConversion = GetListExpressionConversion(listExpression, destination);
-
-        if (listExpressionConversion.exists)
-            return listExpressionConversion;
-
-        if (destination.IsNullableType(out var underlyingDestination)) {
-            var underlyingConversion = GetListExpressionConversion(listExpression, underlyingDestination);
-
-            if (underlyingConversion.exists)
-                return new Conversion(ConversionKind.ImplicitNullable, [underlyingConversion]);
-        }
-
-        return Conversion.None;
-    }
-
-    internal Conversion GetNullptrExpressionConversion(BoundUnconvertedNullptrExpression node, TypeSymbol targetType) {
-        if (targetType.IsPointerOrFunctionPointer())
-            return Conversion.ImplicitNullToPointer;
-
-        return Conversion.None;
-    }
-
-    internal Conversion GetEnumFieldExpressionConversion(
-        BoundUnconvertedImplicitEnumFieldExpression node,
-        TypeSymbol targetType) {
-        if (targetType.StrippedType().IsEnumType()) {
-            var conversion = Conversion.Identity;
-
-            if (targetType.IsNullableType())
-                conversion = new Conversion(ConversionKind.ImplicitNullable, [conversion]);
-
-            return conversion;
-        }
-
-        return Conversion.None;
-    }
-
-    internal static Conversion GetExtendedLiteralExpressionConversion(
-        Binder binder,
-        BoundUnconvertedExtendedLiteralExpression node,
-        TypeSymbol targetType) {
-        var candidates = binder.GetExtendedLiteralCandidates(
-            node,
-            targetType,
-            out var analyzedArguments,
-            out _
-        );
-
-        analyzedArguments?.Free();
-
-        if (!candidates.hasAnyApplicableMember)
-            return Conversion.None;
-
-        if (candidates.succeeded) {
-            var method = candidates.bestResult.member;
-            candidates.Free();
-            return new Conversion(ConversionKind.Implicit, method);
-        }
-
-        candidates.Free();
-        return Conversion.Implicit;
-    }
-
-    internal Conversion GetListExpressionConversion(BoundUnconvertedInitializerList node, TypeSymbol targetType) {
-        var listTypeKind = GetListExpressionTypeKind(targetType, out var elementTypeWithAnnotations);
-        var elementType = elementTypeWithAnnotations?.type;
-
-        switch (listTypeKind) {
-            case ListExpressionTypeKind.None:
-                return Conversion.None;
-        }
-
-        var items = node.items;
-
-        var builder = ArrayBuilder<Conversion>.GetInstance(items.Length);
-
-        foreach (var element in items) {
-            var elementConversion = ClassifyImplicitConversionFromExpression(element, elementType);
-
-            if (!elementConversion.exists) {
-                builder.Free();
-                return Conversion.None;
-            }
-
-            builder.Add(elementConversion);
-        }
-
-        return Conversion.CreateListExpressionConversion(listTypeKind, elementType, builder.ToImmutableAndFree());
-    }
-
-    internal static Conversion FastClassifyConversion(TypeSymbol source, TypeSymbol target) {
-        var conversionKind = Conversion.EasyOut.Classify(source, target);
-
-        if (conversionKind != ConversionKind.ImplicitNullable && conversionKind != ConversionKind.ExplicitNullable)
-            return new Conversion(conversionKind);
-
-        return Conversion.MakeNullableConversion(
-            conversionKind,
-            FastClassifyConversion(source.StrippedType(), target.StrippedType())
-        );
-    }
-
-    internal static bool HasIdentityConversion(TypeSymbol source, TypeSymbol target) {
-        return source.Equals(target, TypeCompareKind.AllIgnoreOptions);
-    }
-
-    internal Conversion ClassifyConversionFromExpression(BoundExpression sourceExpression, TypeSymbol target) {
-        var result = ClassifyImplicitConversionFromExpression(sourceExpression, target);
-
-        if (result.exists || sourceExpression.IsLiteralNull() ||
-            sourceExpression is BoundUnconvertedInitializerList or
-                                BoundUnconvertedImplicitEnumFieldExpression or
-                                BoundUnconvertedNullptrExpression or
-                                BoundUnconvertedExtendedLiteralExpression or
-                                BoundUnconvertedConditionalOperator or
-                                BoundUnconvertedArrayLength or
-                                BoundMethodGroup) {
-            // We tried our best. No further built-in conversions for these cases.
-            return result;
-        }
-
-        return ClassifyExplicitOnlyConversionFromExpression(sourceExpression, target);
-    }
-
-    private Conversion ClassifyExplicitOnlyConversionFromExpression(
-        BoundExpression sourceExpression,
-        TypeSymbol target) {
-        // TODO explicit tuple literal conversion
-        // if (sourceExpression.kind == BoundKind.TupleLiteral) {
-        //     Conversion tupleConversion = ClassifyExplicitTupleLiteralConversion(
-        //         (BoundTupleLiteral)sourceExpression,
-        //         target
-        //     );
-
-        //     if (tupleConversion.exists)
-        //         return tupleConversion;
-        // }
-
-        sourceExpression = Binder.ReduceNumericIfApplicable(target, sourceExpression);
-        var sourceType = sourceExpression.Type();
-
-        if (sourceType is not null) {
-            var fastConversion = FastClassifyConversion(sourceType, target);
-
-            if (fastConversion.exists) {
-                return fastConversion;
-            } else {
-                var conversion = ClassifyExplicitBuiltInConversion(sourceType, target);
-
-                if (conversion.exists)
-                    return conversion;
-            }
-        }
-
-        return GetExplicitUserDefinedConversion(sourceExpression, sourceType, target);
-    }
-
-    private Conversion GetImplicitUserDefinedConversion(
-        BoundExpression sourceExpression,
-        TypeSymbol source,
-        TypeSymbol destination) {
-        var conversionResult = AnalyzeImplicitUserDefinedConversions(
-            sourceExpression,
-            source,
-            destination
-        );
-
-        return new Conversion(conversionResult, isImplicit: true);
-    }
-
-    private Conversion GetImplicitUserDefinedConversion(TypeSymbol source, TypeSymbol destination) {
-        return GetImplicitUserDefinedConversion(sourceExpression: null, source, destination);
-    }
-
-    private Conversion GetExplicitUserDefinedConversion(
-        BoundExpression sourceExpression,
-        TypeSymbol source,
-        TypeSymbol destination) {
-        var conversionResult = AnalyzeExplicitUserDefinedConversions(sourceExpression, source, destination);
-        return new Conversion(conversionResult, isImplicit: false);
-    }
-
-    private Conversion GetExplicitUserDefinedConversion(TypeSymbol source, TypeSymbol destination) {
-        return GetExplicitUserDefinedConversion(sourceExpression: null, source, destination);
-    }
-
-    internal Conversion ClassifyBuiltInConversion(TypeSymbol source, TypeSymbol target) {
-        var conversion = FastClassifyConversion(source, target);
-
-        if (conversion.exists)
-            return conversion;
-
-        return Conversion.Classify(source, target);
-    }
-
-    internal Conversion ClassifyConversionFromType(TypeSymbol source, TypeSymbol target) {
-        var conversion = FastClassifyConversion(source, target);
-
-        if (conversion.exists) {
-            return conversion;
-        } else {
-            conversion = ClassifyImplicitBuiltInConversionSlow(source, target);
-
-            if (Conversion.CollapseConversion(conversion).isImplicit)
-                return conversion;
-        }
-
-        conversion = GetImplicitUserDefinedConversion(source, target);
-
-        if (conversion.exists)
-            return conversion;
-
-        conversion = ClassifyExplicitBuiltInConversion(source, target);
-
-        if (conversion.exists)
-            return conversion;
-
-        return GetExplicitUserDefinedConversion(source, target);
-    }
-
-    private Conversion ClassifyImplicitBuiltInConversionSlow(TypeSymbol source, TypeSymbol target) {
-        var conversion = Conversion.Classify(source, target);
-
-        if (conversion.isImplicit)
-            return conversion;
-
-        if (HasIdentityConversionInternal(source, target))
-            return Conversion.Identity;
-
-        var nullableConversion = ClassifyImplicitNullableConversion(source, target);
-
-        if (nullableConversion.exists)
-            return nullableConversion;
-
-        if (HasImplicitReferenceConversion(source, target))
-            return Conversion.ImplicitReference;
-
-        if (HasBoxingConversion(source, target))
-            return Conversion.AnyBoxing;
-
-        return Conversion.None;
-    }
-
-    private Conversion ClassifyImplicitNullableConversion(TypeSymbol source, TypeSymbol target) {
-        if (target.IsNullableType()) {
-            var underlyingConversion = ClassifyImplicitConversionFromType(source.StrippedType(), target.StrippedType());
-
-            if (underlyingConversion.isIdentity && source.isReferenceType)
-                return underlyingConversion;
-
-            if (underlyingConversion.exists)
-                return new Conversion(ConversionKind.ImplicitNullable, [underlyingConversion]);
-        }
-
-        return Conversion.None;
-    }
-
-    private Conversion ClassifyExplicitNullableConversion(TypeSymbol source, TypeSymbol target) {
-        if (source.IsNullableType() && !target.IsNullableType()) {
-            var underlyingConversion = ClassifyImplicitConversionFromType(source.StrippedType(), target);
-
-            if (underlyingConversion.exists)
-                return new Conversion(ConversionKind.ExplicitNullable, [underlyingConversion]);
-        }
-
-        return Conversion.None;
-    }
-
-    internal bool HasBoxingConversion(TypeSymbol source, TypeSymbol destination) {
-        if ((source.typeKind == TypeKind.TemplateParameter) &&
-            HasImplicitBoxingTemplateParameterConversion((TemplateParameterSymbol)source, destination)) {
-            return true;
-        }
-
-        if (!source.isValueType || !destination.isReferenceType)
-            return false;
-
-        if (source.IsNullableType())
-            return HasBoxingConversion(source.GetNullableUnderlyingType(), destination);
-
-        if (IsBaseClass(source, destination))
-            return true;
-
-        if (HasAnyBaseInterfaceConversion(source, destination))
-            return true;
-
-        return false;
-    }
-
-    private bool HasImplicitBoxingTemplateParameterConversion(TemplateParameterSymbol source, TypeSymbol destination) {
-        if (source.isReferenceType)
-            return false;
-
-        if (source.allowsRefLikeType)
-            return false;
-
-        if (HasImplicitEffectiveBaseConversion(source, destination))
-            return true;
-
-        if (HasImplicitEffectiveInterfaceSetConversion(source, destination))
-            return true;
-
-        if (destination is TemplateParameterSymbol { allowsRefLikeType: false } d && source.DependsOn(d))
-            return true;
-
-        return false;
-    }
-
-    private Conversion ClassifyExplicitBuiltInConversion(TypeSymbol source, TypeSymbol target) {
-        var conversion = Conversion.Classify(source, target);
-
-        if (conversion.exists)
-            return conversion;
-
-        var nullableConversion = ClassifyExplicitNullableConversion(source, target);
-
-        if (nullableConversion.exists)
-            return nullableConversion;
-
-        if (HasExplicitReferenceConversion(source, target))
-            return Conversion.ExplicitReference;
-
-        if (HasUnboxingConversion(source, target))
-            return Conversion.AnyUnboxing;
-
-        return Conversion.None;
-    }
-
-    private bool HasUnboxingConversion(TypeSymbol source, TypeSymbol destination) {
-        if (destination.IsPointerOrFunctionPointer())
-            return false;
-
-        var specialTypeSource = source.specialType;
-
-        if (specialTypeSource == SpecialType.Object || specialTypeSource == SpecialType.ValueType) {
-            if (destination.isValueType && !destination.IsNullableType())
-                return true;
-        }
-
-        if (source.IsInterfaceType() &&
-            destination.isValueType &&
-            !destination.IsNullableType() &&
-            HasBoxingConversion(destination, source)) {
-            return true;
-        }
-
-        if (source.specialType == SpecialType.Enum && destination.IsEnumType())
-            return true;
-
-        if (source.isReferenceType &&
-            destination.IsNullableType() &&
-            HasUnboxingConversion(source, destination.GetNullableUnderlyingType())) {
-            return true;
-        }
-
-        if (HasUnboxingTemplateParameterConversion(source, destination))
-            return true;
-
-        return false;
-    }
-
-    private bool HasUnboxingTemplateParameterConversion(TypeSymbol source, TypeSymbol destination) {
-        var s = source as TemplateParameterSymbol;
-        var t = destination as TemplateParameterSymbol;
-
-        if (s?.allowsRefLikeType == true || t?.allowsRefLikeType == true)
-            return false;
-
-        if (t is not null && !t.isReferenceType) {
-            for (var type = t.effectiveBaseClass; type is not null; type = type.baseType) {
-                if (TypeSymbol.Equals(type, source, TypeCompareKind.ConsiderEverything))
-                    return true;
-            }
-        }
-
-        if (source.IsInterfaceType() && t is not null && !t.isReferenceType)
-            return true;
-
-        if (s is not null && !s.isReferenceType && destination.IsInterfaceType() &&
-            !HasImplicitReferenceTemplateParameterConversion(s, destination)) {
-            return true;
-        }
-
-        if (s is not null && t is not null && !t.isReferenceType && t.DependsOn(s))
-            return true;
-
-        return false;
-    }
-
-    private bool HasExplicitReferenceConversion(TypeSymbol source, TypeSymbol destination) {
-        if (source.specialType == SpecialType.Object) {
-            if (destination.isReferenceType)
-                return true;
-        }
-
-        if (destination.IsClassType() && IsBaseClass(destination, source))
-            return true;
-
-        if (source.IsClassType() && destination.IsInterfaceType() && !source.isSealed &&
-            !HasAnyBaseInterfaceConversion(source, destination)) {
-            return true;
-        }
-
-        if (source.IsInterfaceType() && destination.IsClassType() &&
-            (!destination.isSealed || HasAnyBaseInterfaceConversion(destination, source))) {
-            return true;
-        }
-
-        if (source.IsInterfaceType() && destination.IsInterfaceType() &&
-            !HasImplicitConversionToInterface(source, destination)) {
-            return true;
-        }
-
-        // TODO
-        // if (HasExplicitArrayConversion(source, destination)) {
-        //     return true;
-        // }
-
-        if (HasExplicitReferenceTemplateParameterConversion(source, destination))
-            return true;
-
-        return false;
-    }
-
-    private bool HasExplicitReferenceTemplateParameterConversion(TypeSymbol source, TypeSymbol destination) {
-        var s = source as TemplateParameterSymbol;
-        var t = destination as TemplateParameterSymbol;
-
-        if (s?.allowsRefLikeType == true || t?.allowsRefLikeType == true)
-            return false;
-
-        if (t is not null && t.isReferenceType) {
-            for (var type = t.effectiveBaseClass; type is not null; type = type.baseType) {
-                if (HasIdentityConversionInternal(type, source))
-                    return true;
-            }
-        }
-
-        if (t is not null && source.IsInterfaceType() && t.isReferenceType)
-            return true;
-
-        if (s is not null && s.isReferenceType && destination.IsInterfaceType() &&
-            !HasImplicitReferenceTemplateParameterConversion(s, destination)) {
-            return true;
-        }
-
-        if (s is not null && t is not null && t.isReferenceType && t.DependsOn(s))
-            return true;
-
-        return false;
-    }
-
-    internal Conversion ClassifyImplicitConversionFromType(TypeSymbol source, TypeSymbol target) {
-        var conversion = ClassifyConversionFromType(source, target);
-
-        if (conversion.isImplicit)
-            return conversion;
-
-        return Conversion.None;
-    }
-
-    private Conversion ClassifyImplicitBuiltInConversionFromExpression(
-        BoundExpression sourceExpression,
-        TypeSymbol source,
-        TypeSymbol target) {
-        switch (sourceExpression) {
-            case BoundUnconvertedInitializerList list:
-                var listExpressionConversion = GetImplicitListExpressionConversion(list, target);
-
-                if (listExpressionConversion.exists)
-                    return listExpressionConversion;
-
-                // TODO Eventually we will handle user conversions, but right now if we can't immediately convert this
-                // we won't be able to
-                return listExpressionConversion;
-            case BoundUnconvertedNullptrExpression nullptr:
-                return GetImplicitNullptrExpressionConversion(nullptr, target);
-            case BoundUnconvertedImplicitEnumFieldExpression fieldAccess:
-                return GetImplicitEnumFieldExpressionConversion(fieldAccess, target);
-            case BoundMethodGroup methodGroup:
-                return GetMethodGroupConversion(methodGroup, target);
-            case BoundDefaultLiteral:
-                return Conversion.DefaultLiteral;
-            case BoundUnconvertedObjectCreationExpression:
-                return Conversion.ObjectCreation;
-            case BoundUnconvertedExtendedLiteralExpression extended:
-                return GetImplicitExtendedLiteralExpressionConversion(extended, target);
-            case BoundUnconvertedConditionalOperator conditionalOperator:
-                return GetConditionalExpressionConversion(conditionalOperator, target);
-            case BoundTupleLiteral tuple:
-                return GetImplicitTupleConversion(tuple, target);
-            case BoundUnconvertedArrayLength length:
-                return GetImplicitArrayLengthConversion(length, target);
-            case BoundConditionalAccessExpression conditionalAccess
-                when conditionalAccess.accessExpression is BoundUnconvertedArrayLength length:
-                return GetImplicitArrayLengthConversion(length, target);
-        }
-
-        if (sourceExpression.IsLiteralNull()) {
-            if (target.typeKind == TypeKind.Pointer)
-                return Conversion.ImplicitNullToPointer;
-            else if (target.IsNullableType())
-                return Conversion.NullLiteral;
-            else
-                return Conversion.None;
-        }
-
-        return Conversion.None;
-    }
-
-    internal Conversion ClassifyImplicitConversionFromExpression(BoundExpression sourceExpression, TypeSymbol target) {
-        sourceExpression = Binder.ReduceNumericIfApplicable(target, sourceExpression);
-        var sourceType = sourceExpression.Type();
-
-        if (sourceType is not null && HasIdentityConversionInternal(sourceType, target))
-            return Conversion.Identity;
-
-        var conversion = ClassifyImplicitBuiltInConversionFromExpression(sourceExpression, sourceType, target);
-
-        if (conversion.exists && Conversion.CollapseConversion(conversion).isImplicit)
-            return conversion;
-
-        if (sourceType is not null) {
-            var fastConversion = FastClassifyConversion(sourceType, target);
-
-            if (fastConversion.exists) {
-                if (Conversion.CollapseConversion(fastConversion).isImplicit)
-                    return fastConversion;
-            } else {
-                conversion = ClassifyImplicitBuiltInConversionSlow(sourceType, target);
-
-                if (conversion.exists && Conversion.CollapseConversion(conversion).isImplicit)
-                    return conversion;
-            }
-        }
-        // TODO Handle function types here?
-
-        conversion = GetImplicitUserDefinedConversion(sourceExpression, sourceExpression.Type(), target);
-
-        if (conversion.exists)
-            return conversion;
-
-        return Conversion.None;
-    }
-
-    private Conversion GetImplicitArrayLengthConversion(BoundUnconvertedArrayLength length, TypeSymbol destination) {
-        if (destination.specialType is SpecialType.Int32 or SpecialType.Int64 or SpecialType.Int)
-            return Conversion.Identity;
-
-        var conversion = ClassifyConversionFromType(CorLibrary.GetSpecialType(SpecialType.Int), destination);
-
-        if (conversion.isImplicit)
-            return conversion;
-
-        return Conversion.None;
-    }
-
-    private Conversion GetImplicitTupleConversion(BoundTupleLiteral tuple, TypeSymbol destination) {
-        var conversion = GetTupleConversion(tuple, destination);
-
-        if (conversion.exists)
-            return conversion;
-
-        if (destination.IsNullableType()) {
-            var underlyingConversion = GetTupleConversion(tuple, destination.StrippedType());
-
-            if (underlyingConversion.exists)
-                return new Conversion(ConversionKind.ImplicitNullable, [underlyingConversion]);
-        }
-
-        return Conversion.None;
-    }
-
-    private Conversion GetTupleConversion(BoundTupleLiteral tuple, TypeSymbol destination) {
-        var arguments = tuple.arguments;
-
-        if (!destination.IsTupleTypeOfCardinality(arguments.Length))
-            return Conversion.None;
-
-        var targetElementTypes = destination.tupleElementTypes;
-        var argumentConversions = ArrayBuilder<Conversion>.GetInstance(arguments.Length);
-
-        for (var i = 0; i < arguments.Length; i++) {
-            var argument = arguments[i];
-            var result = ClassifyImplicitConversionFromExpression(argument, targetElementTypes[i].type.type);
-
-            if (!result.exists) {
-                argumentConversions.Free();
-                return Conversion.None;
-            }
-
-            argumentConversions.Add(result);
-        }
-
-        return new Conversion(ConversionKind.ImplicitTupleLiteral, argumentConversions.ToImmutableAndFree());
-    }
-
-    private Conversion GetMethodGroupConversion(BoundMethodGroup source, TypeSymbol destination) {
+    internal override Conversion GetMethodGroupConversion(BoundMethodGroup source, TypeSymbol destination) {
         if (destination.StrippedType().typeKind != TypeKind.Function)
             return Conversion.None;
 
@@ -885,15 +37,17 @@ internal sealed partial class Conversions {
         return conversion;
     }
 
-    private static Conversion ToConversion(
-        OverloadResolutionResult<MethodSymbol> result,
-        MethodGroup methodGroup,
-        int parameterCount) {
-        if (!result.succeeded)
-            return Conversion.None;
-
-        var method = result.bestResult.member;
-        return new Conversion(ConversionKind.MethodGroup, method);
+    private static void GetFunctionArguments(
+        SyntaxNode syntax,
+        AnalyzedArguments analyzedArguments,
+        ImmutableArray<ParameterSymbol> delegateParameters) {
+        foreach (var p in delegateParameters) {
+            var parameter = p;
+            analyzedArguments.arguments.Add(new BoundExpressionOrTypeOrConstant(
+                new BoundParameterExpression(syntax, parameter, null, parameter.type)
+            ));
+            analyzedArguments.refKinds.Add(parameter.refKind);
+        }
     }
 
     private static MethodGroupResolution ResolveMethodGroup(
@@ -960,784 +114,5 @@ internal sealed partial class Conversions {
 
         resolution.Free();
         return hasErrors;
-    }
-
-    private static void GetFunctionArguments(
-        SyntaxNode syntax,
-        AnalyzedArguments analyzedArguments,
-        ImmutableArray<ParameterSymbol> delegateParameters) {
-        foreach (var p in delegateParameters) {
-            var parameter = p;
-            analyzedArguments.arguments.Add(new BoundExpressionOrTypeOrConstant(
-                new BoundParameterExpression(syntax, parameter, null, parameter.type)
-            ));
-            analyzedArguments.refKinds.Add(parameter.refKind);
-        }
-    }
-
-    private Conversion ClassifyStandardImplicitConversion(
-        BoundExpression expression,
-        TypeSymbol source,
-        TypeSymbol target) {
-        // TODO Do we need to use expression here
-        var conversion = Conversion.Classify(source, target);
-
-        if (conversion.isImplicit)
-            return conversion;
-
-        return Conversion.None;
-    }
-
-    private Conversion ClassifyStandardConversion(
-        BoundExpression expression,
-        TypeSymbol source,
-        TypeSymbol target) {
-        // TODO Do we need to use expression here
-        return Conversion.Classify(source, target);
-    }
-
-    private UserDefinedConversionResult AnalyzeExplicitUserDefinedConversions(
-        BoundExpression sourceExpression,
-        TypeSymbol source,
-        TypeSymbol target) {
-        var d = ArrayBuilder<(NamedTypeSymbol participatingType, TemplateParameterSymbol constrainedToType)>
-            .GetInstance();
-
-        ComputeUserDefinedExplicitConversionTypeSet(source, target, d);
-
-        var ubuild = ArrayBuilder<UserDefinedConversionAnalysis>.GetInstance();
-        ComputeApplicableUserDefinedExplicitConversionSet(sourceExpression, source, target, d, ubuild);
-        d.Free();
-        var u = ubuild.ToImmutableAndFree();
-
-        if (u.Length == 0)
-            return UserDefinedConversionResult.NoApplicableOperators(u);
-
-        var sx = MostSpecificSourceTypeForExplicitUserDefinedConversion(u, sourceExpression, source);
-
-        if (sx is null)
-            return UserDefinedConversionResult.NoBestSourceType(u);
-
-        var tx = MostSpecificTargetTypeForExplicitUserDefinedConversion(u, target);
-
-        if (tx is null)
-            return UserDefinedConversionResult.NoBestTargetType(u);
-
-        var best = MostSpecificConversionOperator(sx, tx, u);
-
-        if (best is null)
-            return UserDefinedConversionResult.Ambiguous(u);
-
-        return UserDefinedConversionResult.Valid(u, best.Value);
-    }
-
-    private static void ComputeUserDefinedExplicitConversionTypeSet(
-        TypeSymbol source,
-        TypeSymbol target,
-        ArrayBuilder<(NamedTypeSymbol participatingType, TemplateParameterSymbol constrainedToType)> d) {
-        AddTypesParticipatingInUserDefinedConversion(d, source, includeBaseTypes: true);
-        AddTypesParticipatingInUserDefinedConversion(d, target, includeBaseTypes: true);
-    }
-
-    private void ComputeApplicableUserDefinedExplicitConversionSet(
-        BoundExpression sourceExpression,
-        TypeSymbol source,
-        TypeSymbol target,
-        ArrayBuilder<(NamedTypeSymbol participatingType, TemplateParameterSymbol constrainedToType)> d,
-        ArrayBuilder<UserDefinedConversionAnalysis> u) {
-        var haveInterfaces = false;
-
-        foreach ((var declaringType, var constrainedToTypeOpt) in d) {
-            if (declaringType.isInterface) {
-                haveInterfaces = true;
-            } else {
-                AddCandidatesFromType(null, declaringType, sourceExpression, source, target, u);
-            }
-        }
-
-        if (u.Count == 0 && haveInterfaces) {
-            foreach ((var declaringType, var constrainedToTypeOpt) in d) {
-                if (declaringType.isInterface) {
-                    AddCandidatesFromType(
-                        constrainedToTypeOpt: constrainedToTypeOpt,
-                        declaringType,
-                        sourceExpression,
-                        source,
-                        target,
-                        u
-                    );
-                }
-            }
-        }
-
-        void AddCandidatesFromType(
-            TemplateParameterSymbol constrainedToTypeOpt,
-            NamedTypeSymbol declaringType,
-            BoundExpression sourceExpression,
-            TypeSymbol source,
-            TypeSymbol target,
-            ArrayBuilder<UserDefinedConversionAnalysis> u) {
-            AddUserDefinedConversionsToExplicitCandidateSet(
-                sourceExpression,
-                source,
-                target,
-                u,
-                constrainedToTypeOpt,
-                declaringType,
-                isExplicit: true
-            );
-
-            AddUserDefinedConversionsToExplicitCandidateSet(
-                sourceExpression,
-                source,
-                target,
-                u,
-                constrainedToTypeOpt,
-                declaringType,
-                isExplicit: false
-            );
-        }
-    }
-
-    private void AddUserDefinedConversionsToExplicitCandidateSet(
-        BoundExpression sourceExpression,
-        TypeSymbol source,
-        TypeSymbol target,
-        ArrayBuilder<UserDefinedConversionAnalysis> u,
-        TemplateParameterSymbol constrainedToTypeOpt,
-        NamedTypeSymbol declaringType,
-        bool isExplicit) {
-        if (source is not null && source.IsInterfaceType() || target.IsInterfaceType())
-            return;
-
-        var operators = declaringType.GetOperators(
-            isExplicit ? WellKnownMemberNames.ExplicitConversionName : WellKnownMemberNames.ImplicitConversionName);
-
-        var candidates = ArrayBuilder<MethodSymbol>.GetInstance(operators.Length);
-        candidates.AddRange(operators);
-
-        foreach (var op in candidates) {
-            if (op.returnsVoid || op.parameterCount != 1 || op.returnType.typeKind == TypeKind.Error)
-                continue;
-
-            var convertsFrom = op.GetParameterType(0);
-            var convertsTo = op.returnType;
-            var fromConversion = EncompassingExplicitConversion(sourceExpression, source, convertsFrom);
-            var toConversion = EncompassingExplicitConversion(convertsTo, target);
-
-            if (!fromConversion.exists &&
-                source is not null &&
-                source.IsNullableType() &&
-                EncompassingExplicitConversion(source.GetNullableUnderlyingType(), convertsFrom).exists) {
-                fromConversion = ClassifyBuiltInConversion(source, convertsFrom);
-            }
-
-            if (!toConversion.exists &&
-                target is not null &&
-                target.IsNullableType() &&
-                EncompassingExplicitConversion(convertsTo, target.GetNullableUnderlyingType()).exists) {
-                toConversion = ClassifyBuiltInConversion(convertsTo, target);
-            }
-
-            if (fromConversion.exists && toConversion.exists) {
-                if (source is not null && source.IsNullableType() &&
-                    convertsFrom.IsValidNullableTypeArgument() && target.IsNullableType()) {
-                    var nullableFrom = MakeNullableType(convertsFrom);
-                    var nullableTo = convertsTo.IsValidNullableTypeArgument() ? MakeNullableType(convertsTo) : convertsTo;
-                    var liftedFromConversion = EncompassingExplicitConversion(sourceExpression, source, nullableFrom);
-                    var liftedToConversion = EncompassingExplicitConversion(nullableTo, target);
-
-                    u.Add(UserDefinedConversionAnalysis.Lifted(
-                        constrainedToTypeOpt,
-                        op,
-                        liftedFromConversion,
-                        liftedToConversion,
-                        nullableFrom,
-                        nullableTo
-                    ));
-                } else {
-                    if (target.IsNullableType() && convertsTo.IsValidNullableTypeArgument()) {
-                        convertsTo = MakeNullableType(convertsTo);
-                        toConversion = EncompassingExplicitConversion(convertsTo, target);
-                    }
-
-                    if (source is not null && source.IsNullableType() && convertsFrom.IsValidNullableTypeArgument()) {
-                        convertsFrom = MakeNullableType(convertsFrom);
-                        fromConversion = EncompassingExplicitConversion(convertsFrom, source);
-                    }
-
-                    u.Add(UserDefinedConversionAnalysis.Normal(constrainedToTypeOpt, op, fromConversion, toConversion, convertsFrom, convertsTo));
-                }
-            }
-        }
-
-        candidates.Free();
-    }
-
-    private TypeSymbol MostSpecificSourceTypeForExplicitUserDefinedConversion(
-        ImmutableArray<UserDefinedConversionAnalysis> u,
-        BoundExpression sourceExpression,
-        TypeSymbol source) {
-        if (source is not null) {
-            if (u.Any(static (conv, source)
-                => TypeSymbol.Equals(conv.fromType, source, TypeCompareKind.ConsiderEverything), source)) {
-                return source;
-            }
-
-            Func<UserDefinedConversionAnalysis, bool> isValid =
-                conv => IsEncompassedBy(sourceExpression, source, conv.fromType);
-
-            if (u.Any(isValid)) {
-                var result = MostEncompassedType(u, isValid, conv => conv.fromType);
-                return result;
-            }
-        }
-
-        return MostEncompassingType(u, conv => conv.fromType);
-    }
-
-    private TypeSymbol MostSpecificTargetTypeForExplicitUserDefinedConversion(
-        ImmutableArray<UserDefinedConversionAnalysis> u,
-        TypeSymbol target) {
-        if (u.Any(static (conv, target)
-            => TypeSymbol.Equals(conv.toType, target, TypeCompareKind.ConsiderEverything), target)) {
-            return target;
-        }
-
-        Func<UserDefinedConversionAnalysis, bool> isValid = conv => IsEncompassedBy(conv.toType, target);
-
-        if (u.Any(isValid)) {
-            var result = MostEncompassingType(u, isValid, conv => conv.toType);
-            return result;
-        }
-
-        return MostEncompassedType(u, conv => conv.toType);
-    }
-
-    private Conversion EncompassingExplicitConversion(BoundExpression expr, TypeSymbol a, TypeSymbol b) {
-        return ClassifyStandardConversion(expr, a, b);
-    }
-
-    private Conversion EncompassingExplicitConversion(TypeSymbol a, TypeSymbol b) {
-        return EncompassingExplicitConversion(expr: null, a, b);
-    }
-
-    private UserDefinedConversionResult AnalyzeImplicitUserDefinedConversions(
-        BoundExpression sourceExpression,
-        TypeSymbol source,
-        TypeSymbol target) {
-        var d = ArrayBuilder<(NamedTypeSymbol participatingType, TemplateParameterSymbol constrainedToType)>
-            .GetInstance();
-        ComputeUserDefinedImplicitConversionTypeSet(source, target, d);
-
-        var ubuild = ArrayBuilder<UserDefinedConversionAnalysis>.GetInstance();
-        ComputeApplicableUserDefinedImplicitConversionSet(sourceExpression, source, target, d, ubuild);
-        d.Free();
-        var u = ubuild.ToImmutableAndFree();
-
-        if (u.Length == 0)
-            return UserDefinedConversionResult.NoApplicableOperators(u);
-
-        var sx = MostSpecificSourceTypeForImplicitUserDefinedConversion(u, source);
-
-        if (sx is null)
-            return UserDefinedConversionResult.NoBestSourceType(u);
-
-        var tx = MostSpecificTargetTypeForImplicitUserDefinedConversion(u, target);
-
-        if (tx is null)
-            return UserDefinedConversionResult.NoBestTargetType(u);
-
-        var best = MostSpecificConversionOperator(sx, tx, u);
-
-        return best is null
-            ? UserDefinedConversionResult.Ambiguous(u)
-            : UserDefinedConversionResult.Valid(u, best.Value);
-    }
-
-    private static void ComputeUserDefinedImplicitConversionTypeSet(
-        TypeSymbol s,
-        TypeSymbol t,
-        ArrayBuilder<(NamedTypeSymbol participatingType, TemplateParameterSymbol constrainedToType)> d) {
-        AddTypesParticipatingInUserDefinedConversion(d, s, includeBaseTypes: true);
-        AddTypesParticipatingInUserDefinedConversion(d, t, includeBaseTypes: false);
-    }
-
-    private void ComputeApplicableUserDefinedImplicitConversionSet(
-        BoundExpression sourceExpression,
-        TypeSymbol source,
-        TypeSymbol target,
-        ArrayBuilder<(NamedTypeSymbol participatingType, TemplateParameterSymbol constrainedToType)> d,
-        ArrayBuilder<UserDefinedConversionAnalysis> u,
-        bool allowAnyTarget = false) {
-        if (source is not null && source.IsInterfaceType() || target is not null && target.IsInterfaceType())
-            return;
-
-        var haveInterfaces = false;
-
-        foreach ((var declaringType, var constrainedToTypeOpt) in d) {
-            if (declaringType.isInterface) {
-                haveInterfaces = true;
-            } else {
-                AddCandidatesFromType(
-                    constrainedToTypeOpt: null,
-                    declaringType,
-                    sourceExpression,
-                    source,
-                    target,
-                    u,
-                    allowAnyTarget
-                );
-            }
-        }
-
-        if (u.Count == 0 && haveInterfaces) {
-            foreach ((var declaringType, var constrainedToTypeOpt) in d) {
-                if (declaringType.isInterface) {
-                    AddCandidatesFromType(
-                        constrainedToTypeOpt: constrainedToTypeOpt,
-                        declaringType,
-                        sourceExpression,
-                        source,
-                        target,
-                        u,
-                        allowAnyTarget
-                    );
-                }
-            }
-        }
-
-        void AddCandidatesFromType(
-            TemplateParameterSymbol constrainedToTypeOpt,
-            NamedTypeSymbol declaringType,
-            BoundExpression sourceExpression,
-            TypeSymbol source,
-            TypeSymbol target,
-            ArrayBuilder<UserDefinedConversionAnalysis> u,
-            bool allowAnyTarget) {
-            foreach (var op in declaringType.GetOperators(WellKnownMemberNames.ImplicitConversionName)) {
-                if (op.returnsVoid || op.parameterCount != 1)
-                    continue;
-
-                var convertsFrom = op.GetParameterType(0);
-                var convertsTo = op.returnType;
-                var fromConversion = EncompassingImplicitConversion(sourceExpression, source, convertsFrom);
-                var toConversion = allowAnyTarget
-                    ? Conversion.Identity
-                    : EncompassingImplicitConversion(convertsTo, target);
-
-                if (fromConversion.exists && toConversion.exists) {
-                    if (target is not null && target.IsNullableType() && convertsTo.IsValidNullableTypeArgument()) {
-                        convertsTo = MakeNullableType(convertsTo);
-                        toConversion = allowAnyTarget
-                            ? Conversion.Identity
-                            : EncompassingImplicitConversion(convertsTo, target);
-                    }
-
-                    u.Add(UserDefinedConversionAnalysis.Normal(
-                        constrainedToTypeOpt,
-                        op,
-                        fromConversion,
-                        toConversion,
-                        convertsFrom,
-                        convertsTo
-                    ));
-                } else if (source is not null && source.IsNullableType() && convertsFrom.IsValidNullableTypeArgument() &&
-                      (allowAnyTarget || target.IsNullableType())) {
-                    var nullableFrom = MakeNullableType(convertsFrom);
-                    var nullableTo = convertsTo.IsValidNullableTypeArgument()
-                        ? MakeNullableType(convertsTo)
-                        : convertsTo;
-                    var liftedFromConversion = EncompassingImplicitConversion(sourceExpression, source, nullableFrom);
-                    var liftedToConversion = !allowAnyTarget
-                        ? EncompassingImplicitConversion(nullableTo, target)
-                        : Conversion.Identity;
-
-                    if (liftedFromConversion.exists && liftedToConversion.exists) {
-                        u.Add(UserDefinedConversionAnalysis.Lifted(
-                            constrainedToTypeOpt,
-                            op,
-                            liftedFromConversion,
-                            liftedToConversion,
-                            nullableFrom,
-                            nullableTo
-                        ));
-                    }
-                }
-            }
-        }
-    }
-
-    private TypeSymbol MakeNullableType(TypeSymbol type) {
-        return new TypeWithAnnotations(type).SetIsAnnotated().type;
-    }
-
-    private TypeSymbol MostSpecificSourceTypeForImplicitUserDefinedConversion(
-        ImmutableArray<UserDefinedConversionAnalysis> u,
-        TypeSymbol source) {
-        if (source is not null) {
-            if (u.Any(static (conv, source)
-                => TypeSymbol.Equals(conv.fromType, source, TypeCompareKind.ConsiderEverything), source)) {
-                return source;
-            }
-        }
-
-        return MostEncompassedType(u, conv => conv.fromType);
-    }
-
-    private TypeSymbol MostSpecificTargetTypeForImplicitUserDefinedConversion(
-        ImmutableArray<UserDefinedConversionAnalysis> u,
-        TypeSymbol target) {
-        if (u.Any(static (conv, target)
-            => TypeSymbol.Equals(conv.toType, target, TypeCompareKind.ConsiderEverything), target)) {
-            return target;
-        }
-
-        return MostEncompassingType(u, conv => conv.toType);
-    }
-
-    private static int LiftingCount(UserDefinedConversionAnalysis conv) {
-        var count = 0;
-
-        if (!TypeSymbol.Equals(conv.fromType, conv.@operator.GetParameterType(0), TypeCompareKind.ConsiderEverything))
-            count += 1;
-
-        if (!TypeSymbol.Equals(conv.toType, conv.@operator.returnType, TypeCompareKind.ConsiderEverything))
-            count += 1;
-
-        return count;
-    }
-
-    private static int? MostSpecificConversionOperator(
-        TypeSymbol sx,
-        TypeSymbol tx,
-        ImmutableArray<UserDefinedConversionAnalysis> u) {
-        return MostSpecificConversionOperator(
-            conv => TypeSymbol.Equals(conv.fromType, sx, TypeCompareKind.ConsiderEverything) &&
-            TypeSymbol.Equals(conv.toType, tx, TypeCompareKind.ConsiderEverything), u);
-    }
-
-    private static int? MostSpecificConversionOperator(
-        Func<UserDefinedConversionAnalysis, bool> constraint,
-        ImmutableArray<UserDefinedConversionAnalysis> u) {
-        var bestUnlifted = UniqueIndex(u,
-            conv =>
-            constraint(conv) &&
-            LiftingCount(conv) == 0);
-
-        if (bestUnlifted.kind == BestIndexKind.Best)
-            return bestUnlifted.best;
-        else if (bestUnlifted.kind == BestIndexKind.Ambiguous)
-            return null;
-
-        var bestHalfLifted = UniqueIndex(u,
-            conv =>
-            constraint(conv) &&
-            LiftingCount(conv) == 1);
-
-        if (bestHalfLifted.kind == BestIndexKind.Best)
-            return bestHalfLifted.best;
-        else if (bestHalfLifted.kind == BestIndexKind.Ambiguous)
-            return null;
-
-        var bestFullyLifted = UniqueIndex(u,
-            conv =>
-            constraint(conv) &&
-            LiftingCount(conv) == 2);
-
-        if (bestFullyLifted.kind == BestIndexKind.Best)
-            return bestFullyLifted.best;
-        else if (bestFullyLifted.kind == BestIndexKind.Ambiguous)
-            return null;
-
-        return null;
-    }
-
-    private static BestIndex UniqueIndex<T>(ImmutableArray<T> items, Func<T, bool> predicate) {
-        if (items.IsEmpty)
-            return BestIndex.None();
-
-        int? result = null;
-
-        for (var i = 0; i < items.Length; i++) {
-            if (predicate(items[i])) {
-                if (result is null)
-                    result = i;
-                else
-                    return BestIndex.IsAmbiguous(result.Value, i);
-            }
-        }
-
-        return result is null ? BestIndex.None() : BestIndex.HasBest(result.Value);
-    }
-
-    private bool IsEncompassedBy(BoundExpression aExpr, TypeSymbol a, TypeSymbol b) {
-        return EncompassingImplicitConversion(aExpr, a, b).exists;
-    }
-
-    private bool IsEncompassedBy(TypeSymbol a, TypeSymbol b) {
-        return IsEncompassedBy(aExpr: null, a, b);
-    }
-
-    private Conversion EncompassingImplicitConversion(BoundExpression aExpr, TypeSymbol a, TypeSymbol b) {
-        var result = ClassifyStandardImplicitConversion(aExpr, a, b);
-        return IsEncompassingImplicitConversionKind(result.kind) ? result : Conversion.None;
-    }
-
-    private Conversion EncompassingImplicitConversion(TypeSymbol a, TypeSymbol b) {
-        return EncompassingImplicitConversion(aExpr: null, a, b);
-    }
-
-    private static bool IsEncompassingImplicitConversionKind(ConversionKind kind) {
-        switch (kind) {
-            case ConversionKind.None:
-            case ConversionKind.MethodGroup:
-            case ConversionKind.ConditionalExpression:
-            case ConversionKind.ImplicitEnum:
-            case ConversionKind.ExplicitEnum:
-            case ConversionKind.ImplicitUserDefined:
-            case ConversionKind.ExplicitUserDefined:
-            case ConversionKind.ExplicitNullable:
-            case ConversionKind.ExplicitReference:
-            case ConversionKind.ExplicitNumeric:
-            case ConversionKind.ExplicitPointerToInteger:
-            case ConversionKind.ExplicitIntegerToPointer:
-            case ConversionKind.ExplicitPointerToPointer:
-            case ConversionKind.AnyUnboxing:
-                return false;
-            case ConversionKind.Identity:
-            case ConversionKind.ImplicitNullable:
-            case ConversionKind.ImplicitNumeric:
-            case ConversionKind.ImplicitReference:
-            case ConversionKind.ImplicitNullToPointer:
-            case ConversionKind.ImplicitTupleLiteral:
-            case ConversionKind.AnyBoxing:
-            case ConversionKind.ImplicitConstant:
-            case ConversionKind.NullLiteral:
-            case ConversionKind.DefaultLiteral:
-                return true;
-            // TODO Should remove Implicit as a conversion kind because its vague
-            case ConversionKind.Implicit:
-                return true;
-            default:
-                throw ExceptionUtilities.UnexpectedValue(kind);
-        }
-    }
-
-    private Conversion GetConditionalExpressionConversion(
-        BoundUnconvertedConditionalOperator source,
-        TypeSymbol destination) {
-        var trueConversion = ClassifyImplicitConversionFromExpression(
-            source.trueExpression,
-            destination
-        );
-
-        if (!trueConversion.exists)
-            return Conversion.None;
-
-        var falseConversion = ClassifyImplicitConversionFromExpression(
-            source.falseExpression,
-            destination
-        );
-
-        if (!falseConversion.exists)
-            return Conversion.None;
-
-        return Conversion.MakeConditionalExpression([trueConversion, falseConversion]);
-    }
-
-    private TypeSymbol MostEncompassedType<T>(
-        ImmutableArray<T> items,
-        Func<T, TypeSymbol> extract) {
-        return MostEncompassedType(items, x => true, extract);
-    }
-
-    private TypeSymbol MostEncompassedType<T>(
-        ImmutableArray<T> items,
-        Func<T, bool> valid,
-        Func<T, TypeSymbol> extract) {
-        var best = UniqueBestValidIndex(items, valid,
-            (left, right) => {
-                var leftType = extract(left);
-                var rightType = extract(right);
-
-                if (TypeSymbol.Equals(leftType, rightType, TypeCompareKind.ConsiderEverything))
-                    return BetterResult.Equal;
-
-                var leftWins = IsEncompassedBy(leftType, rightType);
-                var rightWins = IsEncompassedBy(rightType, leftType);
-
-                if (leftWins == rightWins)
-                    return BetterResult.Neither;
-
-                return leftWins ? BetterResult.Left : BetterResult.Right;
-            });
-
-        return best is null ? null : extract(items[best.Value]);
-    }
-
-    private TypeSymbol MostEncompassingType<T>(
-        ImmutableArray<T> items,
-        Func<T, TypeSymbol> extract) {
-        return MostEncompassingType(items, x => true, extract);
-    }
-
-    private TypeSymbol MostEncompassingType<T>(
-        ImmutableArray<T> items,
-        Func<T, bool> valid,
-        Func<T, TypeSymbol> extract) {
-        var best = UniqueBestValidIndex(items, valid,
-            (left, right) => {
-                var leftType = extract(left);
-                var rightType = extract(right);
-
-                if (TypeSymbol.Equals(leftType, rightType, TypeCompareKind.ConsiderEverything))
-                    return BetterResult.Equal;
-
-                var leftWins = IsEncompassedBy(rightType, leftType);
-                var rightWins = IsEncompassedBy(leftType, rightType);
-
-                if (leftWins == rightWins)
-                    return BetterResult.Neither;
-
-                return leftWins ? BetterResult.Left : BetterResult.Right;
-            });
-
-        return best is null ? null : extract(items[best.Value]);
-    }
-
-    private static int? UniqueBestValidIndex<T>(
-        ImmutableArray<T> items,
-        Func<T, bool> valid,
-        Func<T, T, BetterResult> better) {
-        if (items.IsEmpty)
-            return null;
-
-        int? candidateIndex = null;
-        var candidateItem = default(T);
-
-        for (var currentIndex = 0; currentIndex < items.Length; ++currentIndex) {
-            var currentItem = items[currentIndex];
-
-            if (!valid(currentItem))
-                continue;
-
-            if (candidateIndex is null) {
-                candidateIndex = currentIndex;
-                candidateItem = currentItem;
-                continue;
-            }
-
-            var result = better(candidateItem, currentItem);
-
-            if (result == BetterResult.Equal) {
-                continue;
-            } else if (result == BetterResult.Neither) {
-                candidateIndex = null;
-                candidateItem = default;
-            } else if (result == BetterResult.Right) {
-                candidateIndex = currentIndex;
-                candidateItem = currentItem;
-            }
-        }
-
-        if (candidateIndex is null)
-            return null;
-
-        for (var currentIndex = 0; currentIndex < candidateIndex.Value; currentIndex++) {
-            var currentItem = items[currentIndex];
-
-            if (!valid(currentItem))
-                continue;
-
-            var result = better(candidateItem, currentItem);
-
-            if (result != BetterResult.Left && result != BetterResult.Equal)
-                return null;
-        }
-
-        return candidateIndex;
-    }
-
-    private UserDefinedConversionResult AnalyzeImplicitUserDefinedConversionForSwitchGoverningType(
-        TypeSymbol source) {
-        var d = ArrayBuilder<(NamedTypeSymbol participatingType, TemplateParameterSymbol constrainedToType)>.GetInstance();
-        ComputeUserDefinedImplicitConversionTypeSet(source, t: null, d: d);
-
-        var ubuild = ArrayBuilder<UserDefinedConversionAnalysis>.GetInstance();
-        ComputeApplicableUserDefinedImplicitConversionSet(
-            sourceExpression: null,
-            source,
-            target: null,
-            d: d,
-            u: ubuild,
-            allowAnyTarget: true
-        );
-
-        d.Free();
-        var u = ubuild.ToImmutableAndFree();
-
-        // TODO See V6 doc
-        var best = MostSpecificConversionOperator(conv => true, u);
-
-        if (best is null)
-            return UserDefinedConversionResult.NoApplicableOperators(u);
-
-        return UserDefinedConversionResult.Valid(u, best.Value);
-    }
-
-    internal static void AddTypesParticipatingInUserDefinedConversion(
-        ArrayBuilder<(NamedTypeSymbol participatingType, TemplateParameterSymbol constrainedToType)> result,
-        TypeSymbol type,
-        bool includeBaseTypes) {
-        if (type is null)
-            return;
-
-        type = type.StrippedType();
-
-        var excludeExisting = result.Count > 0;
-
-        if (type is TemplateParameterSymbol typeParameter) {
-            var effectiveBaseClass = typeParameter.effectiveBaseClass;
-            AddFromClassOrStruct(result, excludeExisting, effectiveBaseClass, includeBaseTypes);
-        } else {
-            AddFromClassOrStruct(result, excludeExisting, type, includeBaseTypes);
-        }
-
-        static void AddFromClassOrStruct(
-            ArrayBuilder<(NamedTypeSymbol participatingType, TemplateParameterSymbol constrainedToType)> result,
-            bool excludeExisting,
-            TypeSymbol type,
-            bool includeBaseTypes) {
-            if (type.IsClassType() || type.IsStructType()) {
-                var namedType = (NamedTypeSymbol)type;
-                if (!excludeExisting || !HasIdentityConversionToAny(namedType, result)) {
-                    result.Add((namedType, null));
-                }
-            }
-
-            if (!includeBaseTypes)
-                return;
-
-            var t = type.baseType;
-
-            while (t is not null) {
-                if (!excludeExisting || !HasIdentityConversionToAny(t, result))
-                    result.Add((t, null));
-
-                t = t.baseType;
-            }
-        }
-    }
-
-    private static bool HasIdentityConversionToAny(
-        NamedTypeSymbol type,
-        ArrayBuilder<(NamedTypeSymbol participatingType, TemplateParameterSymbol constrainedToType)> targetTypes) {
-        foreach (var (participatingType, _) in targetTypes) {
-            if (HasIdentityConversion(type, participatingType))
-                return true;
-        }
-
-        return false;
     }
 }
