+++ src/Buckle/Compiler/CodeAnalysis/Symbols/Source/SourceMemberFieldSymbolFromDeclarator.cs
@@ -50,9 +50,9 @@ internal partial class SourceMemberFieldSymbolFromDeclarator : SourceMemberField
         return GetTypeAndRefKind(fieldsBeingBound).type;
     }
 
-    internal override void AfterAddingTypeMembersChecks(BelteDiagnosticQueue diagnostics) {
-        type.UnderlyingTemplateTypeOrSelf().CheckAllConstraints(errorLocation, diagnostics);
-        base.AfterAddingTypeMembersChecks(diagnostics);
+    internal override void AfterAddingTypeMembersChecks(ConversionsBase conversions, BelteDiagnosticQueue diagnostics) {
+        type.UnderlyingTemplateTypeOrSelf().CheckAllConstraints(conversions, errorLocation, diagnostics);
+        base.AfterAddingTypeMembersChecks(conversions, diagnostics);
     }
 
     internal bool FieldTypeInferred(ConsList<FieldSymbol> fieldsBeingBound) {
