+++ src/Buckle/Compiler/CompilerStage.cs
@@ -5,7 +5,10 @@ namespace Buckle;
 /// The current step in compilation a source file is.
 /// </summary>
 public enum CompilerStage : byte {
+    /// Source file (.blt)
     Raw,
+
+    /// Other
     Compiled,
     Assembled,
     Finished,
